requirements.txt
streamlit
fpdf
python-docx
Pillow
pandas
